model sync, further artifact-based VTL

+++ Plugins/EASy-Producer/ScenariosTest/testdata/real/IIP-Ecosphere/experiment/expected/Modbus/ApplicationInterfaces/src/main/java/iip/datatypes/ModbusPhoenixEEM.java
@@ -1,6 +1,7 @@
 package iip.datatypes;
 
 import de.iip_ecosphere.platform.support.iip_aas.ConfiguredName;
+import java.math.*;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
@@ -18,7 +19,7 @@ public interface ModbusPhoenixEEM {
       * @return the value
       */
     @JsonIgnore
-    public short getDay();
+    public int getDay();
     
     /**
       * Returns the value of {@code Month}. Month of occurrence of the voltage dip (UTC)
@@ -26,7 +27,7 @@ public interface ModbusPhoenixEEM {
       * @return the value
       */
     @JsonIgnore
-    public short getMonth();
+    public int getMonth();
     
     /**
       * Returns the value of {@code Year}. Year of occurrence of the voltage dip (UTC)
@@ -34,7 +35,7 @@ public interface ModbusPhoenixEEM {
       * @return the value
       */
     @JsonIgnore
-    public short getYear();
+    public int getYear();
     
     /**
       * Returns the value of {@code U1}. Effective value voltage U1
@@ -67,7 +68,7 @@ public interface ModbusPhoenixEEM {
      * @param value the new value
      */
     @JsonIgnore
-    public void setDay(short value);
+    public void setDay(int value);
     
     /**
      * Changes the value of {@code Month}. Month of occurrence of the voltage dip (UTC)
@@ -75,7 +76,7 @@ public interface ModbusPhoenixEEM {
      * @param value the new value
      */
     @JsonIgnore
-    public void setMonth(short value);
+    public void setMonth(int value);
     
     /**
      * Changes the value of {@code Year}. Year of occurrence of the voltage dip (UTC)
@@ -83,7 +84,7 @@ public interface ModbusPhoenixEEM {
      * @param value the new value
      */
     @JsonIgnore
-    public void setYear(short value);
+    public void setYear(int value);
     
     /**
      * Changes the value of {@code U1}. Effective value voltage U1
