+++ Plugins/EASy-Producer/ScenariosTest/testdata/real/IIP-Ecosphere/experiment/expected/Modbus/ApplicationInterfaces/src/main/java/iip/datatypes/ModbusPhoenixEEMImpl.java
@@ -2,6 +2,7 @@ package iip.datatypes;
 
 import iip.datatypes.ModbusPhoenixEEM;
 import de.iip_ecosphere.platform.support.iip_aas.ConfiguredName;
+import java.math.*;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
@@ -14,15 +15,15 @@ public class ModbusPhoenixEEMImpl implements ModbusPhoenixEEM {
  
     @JsonProperty("Day")
     @ConfiguredName("Day")
-    private short Day;
+    private int Day;
     
     @JsonProperty("Month")
     @ConfiguredName("Month")
-    private short Month;
+    private int Month;
     
     @JsonProperty("Year")
     @ConfiguredName("Year")
-    private short Year;
+    private int Year;
     
     @JsonProperty("U1")
     @ConfiguredName("U1")
@@ -59,21 +60,21 @@ public class ModbusPhoenixEEMImpl implements ModbusPhoenixEEM {
 
     @Override
     @JsonIgnore
-    public short getDay() {
+    public int getDay() {
     
         return Day;
     }
     
     @Override
     @JsonIgnore
-    public short getMonth() {
+    public int getMonth() {
     
         return Month;
     }
     
     @Override
     @JsonIgnore
-    public short getYear() {
+    public int getYear() {
     
         return Year;
     }
@@ -102,21 +103,21 @@ public class ModbusPhoenixEEMImpl implements ModbusPhoenixEEM {
     
     @Override
     @JsonIgnore
-    public void setDay(short value) {
+    public void setDay(int value) {
     
         this.Day = value;
     }
     
     @Override
     @JsonIgnore
-    public void setMonth(short value) {
+    public void setMonth(int value) {
     
         this.Month = value;
     }
     
     @Override
     @JsonIgnore
-    public void setYear(short value) {
+    public void setYear(int value) {
     
         this.Year = value;
     }
@@ -146,9 +147,9 @@ public class ModbusPhoenixEEMImpl implements ModbusPhoenixEEM {
     @Override
     public int hashCode() {
         int hc = 0;
-        hc += Short.hashCode(getDay());
-        hc += Short.hashCode(getMonth());
-        hc += Short.hashCode(getYear());
+        hc += Integer.hashCode(getDay());
+        hc += Integer.hashCode(getMonth());
+        hc += Integer.hashCode(getYear());
         hc += Float.hashCode(getU1());
         hc += Float.hashCode(getFrequency());
         hc += Float.hashCode(getI1());
